% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheatsheetR.R
\docType{package}
\name{cheatsheet}
\alias{cheatsheet}
\alias{cheatsheet-package}
\title{cheatsheet: Download R Cheatsheets Locally.}
\description{
A simple package to grab cheatsheets and save them to your local computer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bradlindblad/cheatsheetR}
  \item Report bugs at \url{https://github.com/bradlindblad/cheatsheetR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
