% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cheatsheets.R
\name{get_cheatsheets}
\alias{get_cheatsheets}
\title{get_cheatsheets}
\usage{
get_cheatsheets(
  local_path = ".",
  tidyverse_only = FALSE,
  provider,
  url,
  documented = FALSE
)
}
\arguments{
\item{local_path}{local file path to save the cheatsheets to - string}

\item{tidyverse_only}{only download tidyverse cheatsheets? - logical}

\item{provider}{select particular provider? Defaults to "All" - logical}

\item{url}{the url from which the cheatsheets are retrieved - string}

\item{documented}{provide a table of cheatsheets and providers? - logical}
}
\value{
Conditionally returns a table of cheatsheets; called for side effects.
}
\description{
Download all current cheat sheets as pdf files into your chosen directory
}
\note{
Downloads all pdf cheatsheets available as pdf from a url
}
\examples{
\dontrun{
get_cheatsheets("~/user/Cheatsheets", tidyverse_only = FALSE)
}

}
