% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_translation.R
\name{get_translation}
\alias{get_translation}
\title{get_translation}
\usage{
get_translation(local_path = ".", language = NA)
}
\arguments{
\item{local_path}{local file path to save the cheatsheets to - string}

\item{language}{pick a language from the list obtained by running \code{cheatsheet::list_languages()}}
}
\value{
No return value; called for side effects.
}
\description{
Download all current cheat sheets as pdf files into your chosen directory
}
\examples{
\dontrun{
get_translation("~/user/Cheatsheets", language = "german")
}

}
