% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_cheatsheets.R
\name{list_all_cheatsheets}
\alias{list_all_cheatsheets}
\title{list_all_cheatsheets}
\usage{
list_all_cheatsheets()
}
\description{
List all cheatsheets currently available from https://rstudio.com/resources/cheatsheets/
}
\examples{
\dontrun{
list_all_cheatsheets()
}

}
