% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urls}
\alias{urls}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame containing the cheatsheet provider and github url:
\describe{
\item{Provider}{a descriptive name of the provider}
\item{url}{the github url of the cheatsheet}
...
}
}
\source{
manually recorded repository of cheatsheets
}
\usage{
urls
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
