% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheatsheet.R
\docType{package}
\name{cheatsheet}
\alias{cheatsheet}
\alias{cheatsheet-package}
\title{cheatsheet: Download R Cheatsheets Locally.}
\description{
A simple package to grab cheatsheets and save them to your local computer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bradlindblad.github.io/cheatsheet/}
  \item Report bugs at \url{https://github.com/bradlindblad/cheatsheet/issues}
}

}
\author{
\strong{Maintainer}: Brad Lindblad \email{me@bradlindblad.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
