% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_cheatsheets.R
\name{get_all_cheatsheets}
\alias{get_all_cheatsheets}
\title{get_all_cheatsheets}
\usage{
get_all_cheatsheets(local_path = ".", tidyverse_only = FALSE)
}
\arguments{
\item{local_path}{local file path to save the cheatsheets to - string}

\item{tidyverse_only}{only download tidyverse cheatsheets? - logical}
}
\value{
No return value; called for side effects.
}
\description{
Download all current cheat sheets as pdf files into your chosen directory
}
\note{
Downloads all pdf cheatsheets currently available as pdf from https://rstudio.com/resources/cheatsheets/
}
\examples{
\dontrun{
get_all_cheatsheets("~/user/Cheatsheets", tidyverse_only = FALSE)
}

}
