% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_languages.R
\name{list_languages}
\alias{list_languages}
\title{list_languages}
\usage{
list_languages()
}
\value{
No return value; called for side effects.
}
\description{
View all the languages that can be pulled using \code{cheatsheet::get_translation()}
}
\examples{
\dontrun{
list_languages()
}

}
